feat: post survey progress to progressor
