+++ src/hooks/useProgressorUpdate.ts
@@ -1,13 +1,15 @@
 import { useEffect, useRef } from 'react';
-import { useStore } from '../stores/useStore';
-import type { ProgressorData } from '../types';
+import { useDataStore } from '../stores/useDataStore';
+import { saveProgressor } from '../services/api';
 
 export const useProgressorUpdate = () => {
-  const progressorData = useStore((s) => s.progressorData);
-  const progressorHasChanged = useStore((s) => s.progressorHasChanged);
-  const token = useStore((s) => s.token);
-  const tags = useStore((s) => s.tags);
-  const setProgressorHasChanged = useStore((s) => s.setProgressorHasChanged);
+  const progressorData = useDataStore((s) => s.progressorData);
+  const progressorHasChanged = useDataStore((s) => s.progressorHasChanged);
+  const token = useDataStore((s) => s.token);
+  const tags = useDataStore((s) => s.tags);
+  const setProgressorHasChanged = useDataStore(
+    (s) => s.setProgressorHasChanged
+  );
   const intervalRef = useRef<NodeJS.Timeout | null>(null);
 
   useEffect(() => {
@@ -36,44 +38,4 @@ export const useProgressorUpdate = () => {
     tags?.userId,
     token,
   ]);
-};
-
-const saveProgressor = async (
-  token: string,
-  userId: string,
-  progressorData: ProgressorData
-): Promise<boolean> => {
-  const url = 'https://progressor.usetiful.com/api/save';
-  const headers = {
-    'Content-Type': 'application/json',
-    'x-requested-with': 'XMLHttpRequest',
-    'X-AUTH-TOKEN': token,
-  };
-  const body = JSON.stringify({
-    userId,
-    accountToken: token,
-    data: { ...progressorData, tours: JSON.stringify(progressorData.tours) },
-  });
-
-  try {
-    const response = await fetch(url, {
-      keepalive: true,
-      method: 'POST',
-      headers: headers,
-      body: body,
-    });
-
-    if (!response.ok) {
-      console.error(
-        '=======Error=====>',
-        new Error(`Usetiful: connection error ${response.status}`)
-      );
-      return false;
-    }
-    console.log('Usetiful: Progressor is updated!');
-    return true;
-  } catch (error: any) {
-    console.error('=======Error=====>', error.message);
-    return false;
-  }
 };
